requirements.txt
bleach==2.1.2
certifi==2018.1.18
chardet==3.0.4
cytoolz==0.9.0
decorator==4.2.1
Django==2.0.1
entrypoints==0.2.3
eth-abi==0.5.0
eth-keyfile==0.4.1
eth-keys==0.1.0b4
eth-tester==0.1.0b15
eth-utils==0.7.4
fpdf==1.7.2
html5lib==1.0.1
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lru-dict==1.1.6
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
plotly==2.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py-solc==2.1.0
pycryptodome==3.4.8
Pygments==2.2.0
pysha3==1.0.2
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.4
qtconsole==4.3.1
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
web3==4.0.0b6
webencodings==0.5.1
widgetsnbextension==3.1.3